% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zkcmd.r
\docType{class}
\name{zkcmd-class}
\alias{zkcmd}
\alias{zkcmd-class}
\title{Reference class for sending four-letter commands to zookeeper instances}
\description{
Reference class for sending four-letter commands to zookeeper instances
}
\section{Methods}{

\describe{
\item{\code{available_commands()}}{Show available four-letter commands}

\item{\code{conf()}}{Print details about serving configuration.}

\item{\code{connect(host = "localhost", port = 2181L, timeout = 30L)}}{Connect to a Zookeeper instance.}

\item{\code{cons()}}{List full connection/session details for all clients connected to this server.}

\item{\code{crst()}}{Reset connection/session statistics for all conn}

\item{\code{dirs()}}{Shows the total size of snapshot and log files in bytes}

\item{\code{dump()}}{Lists the outstanding sessions and ephemeral nodes. This only works on the leader.}

\item{\code{envi()}}{Print details about serving environment.}

\item{\code{gtmk()}}{Gets the current trace mask as a 64-bit signed long value in decimal format.}

\item{\code{initialize(..., host = "localhost", port = 2181L, timeout = 30L)}}{Initialize a new Zookeeper commander object.}

\item{\code{isro()}}{Tests if server is running in read-only mode.}

\item{\code{ruok()}}{Tests if server is running in a non-error state. The server will respond with imok if it is running. Otherwise it will not respond at all.}

\item{\code{send_cmd(cmd)}}{Issue a four-letter command. Used internally but can also be used to take advantage of new commands before a package update.}

\item{\code{srst()}}{Reset server statistics.}

\item{\code{srvr()}}{Lists full details for the server.}

\item{\code{stat()}}{Lists brief details for the server and connected clients.}

\item{\code{wchc()}}{Lists detailed information on watches for the server, by session.}

\item{\code{wchp()}}{Lists detailed information on watches for the server, by path.}

\item{\code{wchs()}}{Lists brief information on watches for the server.}
}}
\note{
Currently supports NIO Zookeeper instances.
}
\examples{
\dontrun{
zk <- zkcmd$new()
zk$ruok()
}
}

