% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zkcmd-package.R
\docType{package}
\name{zkcmd}
\alias{zkcmd}
\alias{zkcmd-package}
\title{Tools to Administer 'Zookeeper' Instances with Four-letter Commands}
\description{
'Zookeeper' enables super-easy multi-node configuration management for distributed
computing. The 'Zookeeper' instances themselves can be managed by issuing "four-letter
words" over a TCP socket connection.. Methods are provided to interface with this
administrative API.
}
\author{
Bob Rudis (bob@rud.is)
}

